Split ModelPotential to separate file
