+++ src/general/model_potential.cpp
@@ -26,19 +26,6 @@ namespace helfem {
       throw std::logic_error("Unrecognized model\n");
     }
 
-    ModelPotential::ModelPotential() {
-    }
-
-    ModelPotential::~ModelPotential() {
-    }
-
-    arma::vec ModelPotential::V(const arma::vec & r) const {
-      arma::vec pot(r.n_elem);
-      for(size_t i=0;i<r.n_elem;i++)
-        pot(i)=V(r(i));
-      return pot;
-    }
-
     RadialPotential::RadialPotential(int n_) : n(n_) {
     }
